refactor(moderation): split monolithic page into per-tab sub-pages, add Jackpot section + 5 missing keys + searchable settings
The single /moderation/{guildId} page had grown to 1085 lines across 6
tabs and ~58 config keys, with no way to search and 5 jackpot
eligibility/payout configs (JACKPOT_RTP_MAX_PCT,
JACKPOT_PAYOUT_PCT, JACKPOT_WINNER_COOLDOWN_DAYS,
JACKPOT_ACTIVITY_WINDOW_DAYS, JACKPOT_ACTIVITY_MIN_DAYS) defined in
the backend enum but missing from the UI entirely. Admins couldn't find
the RTP gate that disables jackpots for high-RTP games — and adding it
to one giant page would have made the discoverability problem worse.

Each tab now lives at its own URL under /moderation/{guildId}/<tab>;
the bare URL redirects to /users. The Settings page (formerly Config)
is reorganised into a dedicated Jackpot section gathering all 8
jackpot keys in one place, gets a sticky live-filter search box that
matches both label text and data-key (so "rtp" or "JACKPOT_RTP" both
work, with hash-sync for shareable links), and Anti-autoclicker is
promoted out of "Economy & casino" into its own section.

The navbar is reorganised into User → Play → Economy → Social → Tools
→ Admin groups; the Casino dropdown is folded into Play alongside the
table games and lottery so every game lives under one heading. A new
ModerationTemplateRowsTest guard iterates the entire Configurations
enum and fails CI if any value lacks a UI surface, preventing the
exact failure mode that left JACKPOT_RTP_MAX_PCT invisible — the
mirror direction also catches typos and orphaned rows.
