--- web/src/main/resources/static/js/moderation.js
@@ -1,451 +1,0 @@
-(function () {
-    'use strict';
-
-    const main = document.querySelector('main[data-guild-id]');
-    if (!main) return;
-
-    const guildId = main.dataset.guildId;
-    const isOwner = main.dataset.isOwner === 'true';
-    const actorId = main.dataset.actorId;
-
-    const postJson = window.TobyApi.postJson;
-
-    // --- Tabs ---
-    document.querySelectorAll('.tab-btn').forEach(btn => {
-        btn.addEventListener('click', () => {
-            const tab = btn.dataset.tab;
-            document.querySelectorAll('.tab-btn').forEach(b => {
-                const active = b === btn;
-                b.classList.toggle('active', active);
-                b.setAttribute('aria-selected', active ? 'true' : 'false');
-            });
-            document.querySelectorAll('.tab-panel').forEach(p => {
-                p.classList.toggle('active', p.dataset.panel === tab);
-            });
-        });
-    });
-
-    // --- Users tab: client-side search filter ---
-    (function setupUserSearch() {
-        const input = document.getElementById('user-search');
-        const count = document.getElementById('user-search-count');
-        const usersPanel = document.querySelector('.tab-panel[data-panel="users"]');
-        if (!input || !usersPanel) return;
-
-        // Cache each row's lowercased display name so we don't touch the DOM
-        // per keystroke.
-        const rows = Array.from(usersPanel.querySelectorAll('.mod-table tbody tr[data-member-id]'));
-        const nameByRow = new Map(
-            rows.map(tr => {
-                const nameCell = tr.querySelector('.member-cell span');
-                return [tr, (nameCell?.textContent || '').trim().toLowerCase()];
-            })
-        );
-        const total = rows.length;
-
-        function apply(query) {
-            const q = query.trim().toLowerCase();
-            let shown = 0;
-            rows.forEach(tr => {
-                const hit = q === '' || (nameByRow.get(tr) || '').includes(q);
-                tr.classList.toggle('is-hidden', !hit);
-                if (hit) shown++;
-            });
-            if (count) {
-                count.textContent = q === '' ? '' : shown + ' of ' + total;
-            }
-        }
-
-        input.addEventListener('input', () => apply(input.value));
-        // Esc clears the filter.
-        input.addEventListener('keydown', e => {
-            if (e.key === 'Escape') { input.value = ''; apply(''); }
-        });
-    })();
-
-    // --- Users tab: permission toggles ---
-    document.querySelectorAll('.toggle-btn').forEach(btn => {
-        btn.addEventListener('click', () => {
-            const row = btn.closest('tr');
-            const memberId = row.dataset.memberId;
-            if (memberId === actorId) {
-                toast("You can't change your own permissions.", 'error');
-                return;
-            }
-            if (btn.dataset.ownerOnly === 'true' && !isOwner) {
-                toast('Only the server owner can toggle SUPERUSER.', 'error');
-                return;
-            }
-            btn.disabled = true;
-            postJson('/moderation/' + guildId + '/user/' + memberId + '/permission', {
-                permission: btn.dataset.permission
-            }).then(r => {
-                btn.disabled = false;
-                if (r && r.ok) {
-                    const on = !btn.classList.contains('on');
-                    btn.classList.toggle('on', on);
-                    btn.textContent = on ? 'On' : 'Off';
-                    toast('Permission updated.', 'success');
-                } else {
-                    toast(r?.error || 'Could not update permission.', 'error');
-                }
-            }).catch(() => {
-                btn.disabled = false;
-                toast('Network error.', 'error');
-            });
-        });
-    });
-
-    // --- Users tab: kick ---
-    document.querySelectorAll('.kick-btn').forEach(btn => {
-        btn.addEventListener('click', () => {
-            const row = btn.closest('tr');
-            const memberId = row.dataset.memberId;
-            const name = btn.dataset.name || 'this member';
-            const reason = prompt('Kick ' + name + '? Optional reason:');
-            if (reason === null) return;
-            btn.disabled = true;
-            postJson('/moderation/' + guildId + '/kick', {
-                targetDiscordId: memberId,
-                reason: reason || null
-            }).then(r => {
-                btn.disabled = false;
-                if (r && r.ok) {
-                    toast('Kicked ' + name + '.', 'success');
-                    row.remove();
-                } else {
-                    toast(r?.error || 'Could not kick.', 'error');
-                }
-            }).catch(() => { btn.disabled = false; toast('Network error.', 'error'); });
-        });
-    });
-
-    // --- Config tab ---
-    document.querySelectorAll('.config-row').forEach(row => {
-        const key = row.dataset.key;
-        const input = row.querySelector('input, select');
-        const btn = row.querySelector('.save-config');
-        if (!btn || !input) return;
-        btn.addEventListener('click', () => {
-            const value = (input.value || '').toString();
-            console.debug('[config save]', { key: key, value: value });
-            btn.disabled = true;
-            postJson('/moderation/' + guildId + '/config', { key: key, value: value })
-                .then(r => {
-                    btn.disabled = false;
-                    console.debug('[config save response]', { key: key, response: r });
-                    if (r && r.ok) {
-                        toast('Config saved.', 'success');
-                        // Reflect the saved value as the new default so a refresh
-                        // doesn't look like "nothing happened" even if the read path
-                        // is cached stale somewhere.
-                        if (input.tagName === 'SELECT') {
-                            Array.from(input.options).forEach(o => o.defaultSelected = (o.value === value));
-                        } else {
-                            input.defaultValue = value;
-                        }
-                    } else {
-                        toast(r?.error || 'Could not save config.', 'error');
-                    }
-                })
-                .catch((err) => {
-                    btn.disabled = false;
-                    console.error('[config save error]', { key: key, err: err });
-                    toast('Network error.', 'error');
-                });
-        });
-    });
-
-    // --- Users tab: social credit adjust ---
-    document.querySelectorAll('.sc-apply').forEach(btn => {
-        btn.addEventListener('click', () => {
-            const row = btn.closest('tr');
-            const memberId = row.dataset.memberId;
-            const input = row.querySelector('.sc-adjust input');
-            const delta = parseInt(input.value, 10);
-            if (Number.isNaN(delta) || delta === 0) {
-                toast('Enter a non-zero delta.', 'error');
-                return;
-            }
-            btn.disabled = true;
-            postJson('/moderation/' + guildId + '/user/' + memberId + '/social-credit', { delta: delta })
-                .then(r => {
-                    btn.disabled = false;
-                    if (r && r.ok) {
-                        const scoreCell = row.querySelector('.sc-score');
-                        if (scoreCell) {
-                            const current = parseInt(scoreCell.textContent, 10) || 0;
-                            scoreCell.textContent = String(current + delta);
-                        }
-                        input.value = '0';
-                        toast('Updated social credit.', 'success');
-                    } else {
-                        toast(r?.error || 'Could not adjust.', 'error');
-                    }
-                })
-                .catch(() => { btn.disabled = false; toast('Network error.', 'error'); });
-        });
-    });
-
-    // --- Voice tab: mute/unmute ---
-    document.querySelectorAll('.mute-btn').forEach(btn => {
-        btn.addEventListener('click', () => {
-            const card = btn.closest('.voice-card');
-            const channelId = card.dataset.channelId;
-            const mute = btn.dataset.mute === 'true';
-            btn.disabled = true;
-            postJson('/moderation/' + guildId + '/voice/' + channelId + '/mute', { mute: mute })
-                .then(r => {
-                    btn.disabled = false;
-                    if (r && r.ok) {
-                        const label = mute ? 'muted' : 'unmuted';
-                        const n = (r.changed || []).length;
-                        toast(label + ' ' + n + ' member' + (n === 1 ? '' : 's') + '.', 'success');
-                        if (r.skipped && r.skipped.length) toast('Skipped: ' + r.skipped.join(', '), 'info');
-                    } else {
-                        toast(r?.error || 'Could not change mute.', 'error');
-                    }
-                })
-                .catch(() => { btn.disabled = false; toast('Network error.', 'error'); });
-        });
-    });
-
-    // --- Voice tab: move form ---
-    const moveForm = document.querySelector('.move-form');
-    if (moveForm) {
-        moveForm.addEventListener('submit', e => {
-            e.preventDefault();
-            const targetChannelId = moveForm.elements.targetChannelId.value;
-            const memberIds = Array.from(moveForm.elements.memberIds.selectedOptions).map(o => o.value);
-            if (!targetChannelId || memberIds.length === 0) {
-                toast('Pick a destination and at least one member.', 'error');
-                return;
-            }
-            const submitBtn = moveForm.querySelector('button[type="submit"]');
-            submitBtn.disabled = true;
-            postJson('/moderation/' + guildId + '/voice/move', {
-                targetChannelId: targetChannelId,
-                memberIds: memberIds
-            }).then(r => {
-                submitBtn.disabled = false;
-                if (r && r.ok) {
-                    const n = (r.moved || []).length;
-                    toast('Moved ' + n + ' member' + (n === 1 ? '' : 's') + '.', 'success');
-                    if (r.skipped && r.skipped.length) toast('Skipped: ' + r.skipped.join(', '), 'info');
-                } else {
-                    toast(r?.error || 'Could not move.', 'error');
-                }
-            }).catch(() => { submitBtn.disabled = false; toast('Network error.', 'error'); });
-        });
-    }
-
-    // --- Poll tab ---
-    const pollForm = document.querySelector('.poll-form');
-    if (pollForm) {
-        pollForm.addEventListener('submit', e => {
-            e.preventDefault();
-            const channelId = pollForm.elements.channelId.value;
-            const question = pollForm.elements.question.value.trim();
-            const options = Array.from(pollForm.querySelectorAll('.poll-option'))
-                .map(i => i.value.trim())
-                .filter(v => v.length > 0);
-            if (!channelId || !question || options.length < 2) {
-                toast('Need a channel, a question, and at least 2 options.', 'error');
-                return;
-            }
-            const submitBtn = pollForm.querySelector('button[type="submit"]');
-            submitBtn.disabled = true;
-            postJson('/moderation/' + guildId + '/poll', {
-                channelId: channelId,
-                question: question,
-                options: options
-            }).then(r => {
-                submitBtn.disabled = false;
-                if (r && r.ok) {
-                    toast('Poll posted.', 'success');
-                    pollForm.reset();
-                } else {
-                    toast(r?.error || 'Could not post poll.', 'error');
-                }
-            }).catch(() => { submitBtn.disabled = false; toast('Network error.', 'error'); });
-        });
-    }
-
-    // --- Casino tab ---
-    const jackpotPoolEl = document.getElementById('jackpot-pool');
-    const jackpotResetBtn = document.getElementById('jackpot-reset');
-    if (jackpotResetBtn) {
-        jackpotResetBtn.addEventListener('click', () => {
-            const current = Number(jackpotPoolEl?.textContent || '0');
-            if (current === 0) {
-                toast('Pool is already empty.', 'info');
-                return;
-            }
-            if (!confirm('Reset the jackpot pool to 0? This drains ' + current + ' credits.')) return;
-            jackpotResetBtn.disabled = true;
-            postJson('/moderation/' + guildId + '/jackpot/reset', {}).then(r => {
-                jackpotResetBtn.disabled = false;
-                if (r && r.ok) {
-                    if (jackpotPoolEl) jackpotPoolEl.textContent = String(r.newPool ?? 0);
-                    toast('Jackpot reset. Drained ' + (r.drained ?? 0) + ' credits.', 'success');
-                } else {
-                    toast(r?.error || 'Could not reset jackpot.', 'error');
-                }
-            }).catch(() => { jackpotResetBtn.disabled = false; toast('Network error.', 'error'); });
-        });
-    }
-    const jackpotRefundForm = document.querySelector('.jackpot-refund-form');
-    if (jackpotRefundForm) {
-        const sourceSelect = jackpotRefundForm.elements.sourceDiscordId;
-        const balanceHint = jackpotRefundForm.querySelector('.jackpot-refund-balance');
-        const balanceValue = jackpotRefundForm.querySelector('.jackpot-refund-balance-value');
-        const updateBalanceHint = () => {
-            const opt = sourceSelect.selectedOptions[0];
-            const credit = opt && opt.dataset.credit;
-            if (credit != null && opt.value) {
-                balanceValue.textContent = credit;
-                balanceHint.hidden = false;
-            } else {
-                balanceHint.hidden = true;
-            }
-        };
-        sourceSelect.addEventListener('change', updateBalanceHint);
-        jackpotRefundForm.addEventListener('submit', e => {
-            e.preventDefault();
-            const sourceDiscordId = sourceSelect.value;
-            const amount = Number(jackpotRefundForm.elements.amount.value);
-            if (!/^[0-9]+$/.test(sourceDiscordId) || !Number.isFinite(amount) || amount <= 0) {
-                toast('Select a member and enter a positive amount.', 'error');
-                return;
-            }
-            const selectedOpt = sourceSelect.selectedOptions[0];
-            const memberLabel = selectedOpt ? (selectedOpt.textContent.split(' — ')[0] || sourceDiscordId) : sourceDiscordId;
-            if (!confirm('Refund ' + amount + ' credits from ' + memberLabel + ' into the jackpot pool?')) return;
-            const submitBtn = jackpotRefundForm.querySelector('button[type="submit"]');
-            submitBtn.disabled = true;
-            postJson('/moderation/' + guildId + '/jackpot/refund', {
-                sourceDiscordId: sourceDiscordId,
-                amount: amount
-            }).then(r => {
-                submitBtn.disabled = false;
-                if (r && r.ok) {
-                    if (jackpotPoolEl) jackpotPoolEl.textContent = String(r.newPool ?? 0);
-                    if (selectedOpt && r.newSourceBalance != null) {
-                        selectedOpt.dataset.credit = String(r.newSourceBalance);
-                        const name = selectedOpt.textContent.split(' — ')[0];
-                        selectedOpt.textContent = name + ' — ' + r.newSourceBalance + ' credits';
-                    }
-                    toast(
-                        'Refunded ' + (r.drained ?? 0) + ' credits. New pool: ' + (r.newPool ?? 0) + '.',
-                        'success'
-                    );
-                    jackpotRefundForm.reset();
-                    updateBalanceHint();
-                } else {
-                    toast(r?.error || 'Could not refund to jackpot.', 'error');
-                }
-            }).catch(() => { submitBtn.disabled = false; toast('Network error.', 'error'); });
-        });
-    }
-
-    // ==========================================================
-    // Daily lottery — force-draw button
-    // ==========================================================
-    const lotteryForceDrawBtn = document.getElementById('lottery-force-draw');
-    if (lotteryForceDrawBtn) {
-        lotteryForceDrawBtn.addEventListener('click', () => {
-            if (!confirm(
-                'Force the daily lottery cycle to run now? This closes any open draw ' +
-                'and pays prizes by match tier, then opens a fresh draw seeded from the jackpot.'
-            )) return;
-            lotteryForceDrawBtn.disabled = true;
-            postJson('/moderation/' + guildId + '/lottery/draw', {}).then(r => {
-                lotteryForceDrawBtn.disabled = false;
-                if (r && r.ok) {
-                    let msg = 'Lottery cycle ran.';
-                    if (r.drewPrior) {
-                        if (r.priorBelowMinBuyers) {
-                            msg += ' Prior draw cancelled — only ' + (r.priorBuyersHave ?? 0) +
-                                ' distinct buyer(s), need ' + (r.priorBuyersNeed ?? 0) +
-                                '. Refunded buyer(s); seed returned to jackpot.';
-                        } else {
-                            msg += ' Prior draw paid ' + (r.priorTotalPaid ?? 0) +
-                                ' credits (' + (r.priorRolledBack ?? 0) + ' rolled back to jackpot).';
-                        }
-                    }
-                    if (r.openedNew) {
-                        msg += ' New draw seeded with ' + (r.newSeeded ?? 0) + ' credits.';
-                    }
-                    toast(msg, 'success');
-                } else {
-                    toast(r?.error || 'Could not run the lottery cycle.', 'error');
-                }
-            }).catch(() => {
-                lotteryForceDrawBtn.disabled = false;
-                toast('Network error.', 'error');
-            });
-        });
-    }
-
-    // ==========================================================
-    // Generic "Create read-only channel" forms — wires up every
-    // .config-create-channel-form on the page. The form's
-    // `data-target-config` attribute (LOTTERY_CHANNEL,
-    // LEADERBOARD_CHANNEL, etc.) tells the server which config to
-    // auto-set after creating the channel. Page reloads on success
-    // so the new channel appears in the dropdown above with the
-    // right `selected` — cheaper than splicing an <option> in by
-    // hand and re-syncing two places.
-    // ==========================================================
-    document.querySelectorAll('.config-create-channel-form').forEach(form => {
-        const targetConfig = form.dataset.targetConfig;
-        const nameInput = form.querySelector('input[name="name"]');
-        const parentSelect = form.querySelector('select[name="parentCategoryId"]');
-        const newCategoryInput = form.querySelector('input[name="newCategoryName"]');
-        const btn = form.querySelector('.config-create-channel-btn');
-        if (!targetConfig || !nameInput || !btn) return;
-        form.addEventListener('submit', (e) => {
-            e.preventDefault();
-            const name = (nameInput.value || '').trim();
-            if (!name) {
-                toast('Channel name required.', 'error');
-                return;
-            }
-            const parentCategoryId = (parentSelect?.value || '').trim();
-            const newCategoryName = (newCategoryInput?.value || '').trim();
-            // Build a confirm-message tail that reflects what'll happen
-            // server-side. The new-category path takes precedence over
-            // the dropdown so the user isn't surprised by their own input.
-            let categoryNote = '';
-            if (newCategoryName) {
-                categoryNote = ' under a new category "' + newCategoryName + '"';
-            } else if (parentCategoryId) {
-                const parentText = parentSelect.options[parentSelect.selectedIndex]?.text;
-                if (parentText) categoryNote = ' under category "' + parentText + '"';
-            }
-            if (!confirm(
-                'Create a new "#' + name + '" channel' + categoryNote + '? ' +
-                'It will be read-only for @everyone and post-only for TobyBot. ' +
-                'The ' + targetConfig + ' config will be set to this new channel.'
-            )) return;
-            btn.disabled = true;
-            postJson('/moderation/' + guildId + '/channel/create-read-only', {
-                name: name,
-                targetConfig: targetConfig,
-                parentCategoryId: parentCategoryId || null,
-                newCategoryName: newCategoryName || null
-            }).then(r => {
-                btn.disabled = false;
-                if (r && r.ok) {
-                    toast('Created #' + r.channelName + '. Reloading…', 'success');
-                    setTimeout(() => window.location.reload(), 700);
-                } else {
-                    toast(r?.error || 'Could not create channel.', 'error');
-                }
-            }).catch(() => {
-                btn.disabled = false;
-                toast('Network error.', 'error');
-            });
-        });
-    });
-})();
