+++ web/src/main/resources/templates/moderation/voice.html
@@ -1,0 +1,66 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org" lang="en">
+<head th:replace="~{fragments/head :: head('TobyBot - Moderation: Voice', '/css/moderation.css')}"></head>
+<body>
+<a href="#main" class="skip-link">Skip to content</a>
+<div th:replace="~{fragments/navbar :: navbar}"></div>
+
+<main id="main" class="container"
+      th:attr="data-guild-id=${overview.guildId},
+               data-actor-id=${actorDiscordId},
+               data-is-owner=${isOwner}">
+
+    <div th:replace="~{fragments/moderationHeader :: moderationHeader(${overview}, ${isOwner}, 'voice')}"></div>
+
+    <section class="mod-panel" data-panel="voice">
+        <p class="muted mb-4">
+            Pick a voice channel to mass-mute, unmute, or move its occupants.
+            Requires the bot and you to have the Mute Members / Move Members
+            permissions.
+        </p>
+        <div class="voice-grid">
+            <div class="voice-card" th:each="vc : ${overview.voiceChannels}"
+                 th:attr="data-channel-id=${vc.id}">
+                <header>
+                    <span class="vc-name" th:text="${vc.name}">#channel</span>
+                    <span class="vc-count"
+                          th:text="${#lists.size(vc.memberIds)} + ' in channel'">0 in channel</span>
+                </header>
+                <div class="voice-actions">
+                    <button type="button" class="btn mute-btn" data-mute="true">Shh</button>
+                    <button type="button" class="btn mute-btn" data-mute="false">Talk</button>
+                </div>
+            </div>
+        </div>
+
+        <h2 class="mt-5">Move members</h2>
+        <form class="move-form">
+            <label>
+                Destination
+                <select name="targetChannelId" required>
+                    <option value="">&mdash; choose &mdash;</option>
+                    <option th:each="vc : ${overview.voiceChannels}"
+                            th:value="${vc.id}"
+                            th:text="${vc.name}"></option>
+                </select>
+            </label>
+            <label>
+                Members (ctrl+click for multi)
+                <select name="memberIds" multiple size="6" required>
+                    <option th:each="m : ${overview.members}"
+                            th:if="${!m.isBot}"
+                            th:value="${m.id}"
+                            th:text="${m.name}"></option>
+                </select>
+            </label>
+            <button type="submit" class="btn">Move</button>
+        </form>
+    </section>
+</main>
+
+<script th:src="@{/js/api.js}"></script>
+<script th:src="@{/js/toasts.js}"></script>
+<script th:src="@{/js/moderationCommon.js}"></script>
+<script th:src="@{/js/moderation/voice.js}"></script>
+</body>
+</html>
